\name{attachFileDscr}
\alias{attachFileDscr}
\title{Attach file_dscr from another DDI object}
\usage{
  attachFileDscr(master, attachment)
}
\arguments{
  \item{master}{This is the main object, which will be
  extended.}

  \item{attachment}{This object will be attached to the
  master object.}
}
\description{
  Attach file_dscr from another DDI object
}

