\name{ddi_code_book}
\alias{ddi_code_book}
\title{Generate a ddi code book.}
\usage{
ddi_code_book(id = NULL, doc_dscr = NULL, stdy_dscr = NULL,
  file_dscr = NULL)
}
\arguments{
  \item{id}{Variable name.}

  \item{labl}{Human readable label.}

  \item{intrvl}{Variable format.}

  \item{catgry}{data.frame with categories.}

  \item{sum_stat}{data.frame with summary statistics.}
}
\description{
Generate a ddi code book.
}

