\name{ddi_file_dscr}
\alias{ddi_file_dscr}
\title{Generate a ddi code book.}
\usage{
ddi_file_dscr(id, name = NULL, labl = NULL, format = NULL,
  data_dscr = NULL)
}
\arguments{
  \item{id}{Variable name.}

  \item{labl}{Human readable label.}

  \item{intrvl}{Variable format.}

  \item{catgry}{data.frame with categories.}

  \item{sum_stat}{data.frame with summary statistics.}
}
\description{
Generate a ddi code book.
}

