\name{ddi_var}
\alias{ddi_var}
\title{Generate varDscr}
\usage{
ddi_var(id, labl, intrvl = NULL, catgry = NULL, sum_stat = NULL)
}
\arguments{
  \item{id}{Variable name.}

  \item{labl}{Human readable label.}

  \item{intrvl}{Variable format.}

  \item{catgry}{data.frame with categories.}

  \item{sum_stat}{data.frame with summary statistics.}
}
\description{
Generate varDscr
}

