\name{dir2ddi}
\alias{dir2ddi}
\title{Import multiple files from one directory}
\usage{
  dir2ddi(path, file_type = "all", multicore = TRUE)
}
\arguments{
  \item{path}{Path to files}

  \item{file_type}{Accepted filetypes ("dta", "sav", "csv",
  or "all" for the for the previous three)}

  \item{multicore}{Use multicore functionallity}
}
\description{
  Import multiple files from one directory
}

