\name{jstat.labeled_numeric}
\alias{jstat.labeled_numeric}
\title{Calculate frequencies for character objects}
\usage{
jstat.labeled_numeric(variable, time = NULL)
}
\arguments{
  \item{variable}{variable-object}

  \item{time}{Time variable if long data}
}
\description{
Calculate frequencies for character objects
}

