\name{ddi_var}
\alias{ddi_var}
\title{Generate varDscr}
\usage{
  ddi_var(id, labl, var_format, catgry = NULL,
    sum_stat = NULL)
}
\arguments{
  \item{id}{Variable name.}

  \item{labl}{Human readable label.}

  \item{var_format}{Variable format.}

  \item{catgry}{data.frame with categories.}

  \item{sum_stat}{data.frame with summary statistics.}
}
\description{
  Generate varDscr
}

