\name{stata2ddi}
\alias{stata2ddi}
\title{Generate a DDI object from a Stata file}
\usage{
  stata2ddi(filename, data_name, ...)
}
\arguments{
  \item{filename}{Stata file}
  \item{data_name}{The dataset's name}
  \item{data_label}{Label}
  \item{missing.codes}{Vecor with specific missing codes}
  \item{keep_data}{Keep original data in the DDI object (TRUE) or not (FALSE)}
}
